FIX: remove some size_t/unsigned conversion warnings on Windows in SofaFramework, some remains due to the Index type of BaseVector.
Former-commit-id: 525f05d9344dc29d9f64e4490bcc44df6a3d535f

+++ framework/sofa/core/objectmodel/Base.cpp
@@ -459,8 +459,8 @@ bool Base::parseField( const std::string& attribute, const std::string& value)
             ok = false;
         }
         sout << "Link " << linkVec[l]->getName() << " = " << linkVec[l]->getValueString() << sendl;
-        unsigned int s = linkVec[l]->getSize();
-        for (unsigned int i=0; i<s; ++i)
+        size_t s = linkVec[l]->getSize();
+        for (unsigned int i=0; i<(unsigned int)s; ++i)
         {
             sout  << "  " << linkVec[l]->getLinkedPath(i) << " = ";
             Base* b = linkVec[l]->getLinkedBase(i);
